refactor: GameService에서 모든 로직 수행
